refactor post component
